requirements.txt
-i https://pypi.org/simple
certifi==2018.4.16
chardet==3.0.4
decorator==4.3.0
dnspython==1.15.0
idna==2.6
ipaddress==1.0.22
requests==2.18.4
six==1.11.0
urllib3==1.22
validators==0.12.2
# dev
astroid==1.6.4
attrs==18.1.0
autopep8==1.3.5
coverage==4.5.1
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
more-itertools==4.2.0
pluggy==0.6.0
py==1.5.3
pycodestyle==2.4.0
pylint==1.9.1
pytest-cov==2.5.1
pytest==3.6.1
six==1.11.0
wrapt==1.10.11
